Prevent unexpected results when people 'forget' to cast the DateBuilder.
